requirements.txt
deepchem
pandas
numpy
rdkit-pypi
tensorflow
jax
torch
torch-geometric
catboost
